Deep_learning_projects/Cat_dog_classification/deployed/requirements.txt
gradio
tensorflow
opencv-python
numpy
Pillow